refactor copycommentembedcode into own component
