+++ src/core/client/stream/tabs/Comments/Comment/ModerationDropdown/CopyCommentEmbedCodeContainer.css
@@ -1,0 +1,15 @@
+.label {
+  color: var(--palette-text-500);
+  font-family: var(--font-family-primary);
+  font-weight: var(--font-weight-primary-regular);
+  font-size: var(--font-size-3);
+  line-height: 1;
+
+  text-align: left;
+}
+
+.embedCodeCopied {
+  background-color: var(--palette-success-100);
+  color: var(--palette-success-500);
+  font-weight: var(--font-weight-primary-semi-bold);
+}
