+++ src/core/client/stream/tabs/Comments/Comment/ModerationDropdown/ModerationActionsContainer.tsx
@@ -1,17 +1,9 @@
 import { Localized } from "@fluent/react/compat";
 import cn from "classnames";
-import React, {
-  FunctionComponent,
-  useCallback,
-  useMemo,
-  useState,
-} from "react";
-import CopyToClipboard from "react-copy-to-clipboard";
+import React, { FunctionComponent, useCallback, useMemo } from "react";
 import { graphql } from "react-relay";
 
-import { sanitizeAndFindSpoilerAndSarcasmTags } from "coral-common/helpers/sanitize";
 import { useModerationLink } from "coral-framework/hooks";
-import { useCoralContext } from "coral-framework/lib/bootstrap";
 import { useViewerEvent } from "coral-framework/lib/events";
 import {
   useLocal,
@@ -36,6 +28,7 @@ import RejectCommentMutation from "./RejectCommentMutation";
 import UnfeatureCommentMutation from "./UnfeatureCommentMutation";
 
 import styles from "./ModerationActionsContainer.css";
+import CopyCommentEmbedCodeContainer from "./CopyCommentEmbedCodeContainer";
 
 interface Props {
   comment: ModerationActionsContainer_comment;
@@ -61,8 +54,6 @@ const ModerationActionsContainer: FunctionComponent<Props> = ({
       accessToken
     }
   `);
-  const { window } = useCoralContext();
-  const [embedCodeCopied, setEmbedCodeCopied] = useState(false);
 
   const emitGotoModerationEvent = useViewerEvent(GotoModerationEvent);
   const approve = useMutation(ApproveCommentMutation);
@@ -147,80 +138,10 @@ const ModerationActionsContainer: FunctionComponent<Props> = ({
       : comment.author.id !== viewer.id;
   const isQA = story.settings.mode === GQLSTORY_MODE.QA;
 
-  const handleCopyEmbedCode = useCallback(() => {
-    setEmbedCodeCopied(true);
-  }, [setEmbedCodeCopied]);
-
-  function transform(transformWindow: Window, source: string | Node) {
-    // Sanitize source.
-    const [sanitized, spoilerTags, sarcasmTags, blockquoteTags] =
-      sanitizeAndFindSpoilerAndSarcasmTags(transformWindow, source);
-
-    // Attach event handlers to spoiler tags.
-    spoilerTags.forEach((node) => {
-      node.setAttribute(
-        "onclick",
-        "{this.removeAttribute('style');this.removeAttribute('role');this.removeAttribute('title');this.removeAttribute('onclick');}"
-      );
-      node.setAttribute("role", "button");
-      node.setAttribute("style", "background-color: #14171A; color: #14171A;");
-      node.setAttribute("title", "Reveal spoiler");
-      node.innerHTML = `<span aria-hidden="true">${node.innerHTML}</span>`;
-    });
-
-    sarcasmTags.forEach((node) => {
-      node.setAttribute("style", "font-family: monospace;");
-    });
-
-    blockquoteTags.forEach((node) => {
-      node.setAttribute(
-        "style",
-        "background-color: #EAEFF0; border-radius: 3px; margin: 0.5rem 0 0.5rem 0.2rem; padding: 0.2rem;"
-      );
-    });
-
-    // Return results.
-    return sanitized.innerHTML;
-  }
-
   let showCopyCommentEmbed = false;
-  let sanitizedBody;
-  let mediaUrl = null;
   if (comment.body) {
     showCopyCommentEmbed = true;
-    sanitizedBody = transform(window, comment.body);
   }
-  if (comment.revision?.media) {
-    switch (comment.revision.media.__typename) {
-      case "YouTubeMedia":
-        mediaUrl = comment.revision.media.url;
-        break;
-      case "GiphyMedia":
-        mediaUrl = comment.revision.media.url;
-        break;
-      case "TwitterMedia":
-        mediaUrl = comment.revision.media.url;
-        break;
-      case "ExternalMedia":
-        mediaUrl = comment.revision.media.url;
-        break;
-      case "%other":
-        break;
-    }
-  }
-  const embedCode = `<div class="coral-comment-embed" style="background-color: #f4f7f7; padding: 8px;" data-commentID=${
-    comment.id
-  } data-allowReplies="${
-    settings.embeddedComments?.allowReplies ?? true
-  }" data-reactionLabel="${
-    settings.reaction.label
-  }"><div style="margin-bottom: 8px;">${
-    comment.author?.username
-  }</div><div>${sanitizedBody}${
-    mediaUrl
-      ? `<div><br><a href="${mediaUrl}" target="_blank" rel="noopener noreferrer ugc">${mediaUrl}</a></div>`
-      : ""
-  }</div></div>`;
 
   return (
     <>
@@ -387,34 +308,7 @@ const ModerationActionsContainer: FunctionComponent<Props> = ({
         </DropdownButton>
       </Localized>
       {showCopyCommentEmbed && (
-        <>
-          <DropdownDivider />
-          <CopyToClipboard text={embedCode} onCopy={handleCopyEmbedCode}>
-            {embedCodeCopied ? (
-              <DropdownButton
-                className={cn(styles.label, styles.embedCodeCopied)}
-                icon={
-                  <Icon color="success" size="md">
-                    check_circle_outline
-                  </Icon>
-                }
-              >
-                <Localized id="comments-moderationDropdown-embedCodeCopied">
-                  <span>Code copied</span>
-                </Localized>
-              </DropdownButton>
-            ) : (
-              <DropdownButton
-                className={styles.label}
-                icon={<Icon size="md">code</Icon>}
-              >
-                <Localized id="comments-moderationDropdown-embedCode">
-                  <span>Embed comment</span>
-                </Localized>
-              </DropdownButton>
-            )}
-          </CopyToClipboard>
-        </>
+        <CopyCommentEmbedCodeContainer comment={comment} settings={settings} />
       )}
     </>
   );
@@ -431,26 +325,12 @@ const enhanced = withFragmentContainer<Props>({
       body
       revision {
         id
-        media {
-          __typename
-          ... on GiphyMedia {
-            url
-          }
-          ... on TwitterMedia {
-            url
-          }
-          ... on YouTubeMedia {
-            url
-          }
-          ... on ExternalMedia {
-            url
-          }
-        }
       }
       status
       tags {
         code
       }
+      ...CopyCommentEmbedCodeContainer_comment
     }
   `,
   settings: graphql`
@@ -465,12 +345,7 @@ const enhanced = withFragmentContainer<Props>({
           }
         }
       }
-      embeddedComments {
-        allowReplies
-      }
-      reaction {
-        label
-      }
+      ...CopyCommentEmbedCodeContainer_settings
     }
   `,
   story: graphql`
